mlapi/pyproject.toml
[project]
name = "mlapi"
version = "0.1.0"
description = "Digital Coach's Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "assemblyai>=0.40.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "nltk>=3.9.1",
    "pytest>=8.3.5",
    "redis>=5.2.1",
    "rq>=2.3.2",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
    "python-dotenv>=1.1.0",
    "pydantic-settings>=2.9.1",
    "firebase-admin>=7.1.0",
    "openai>=2.29.0",
    "rq-dashboard-fast>=0.8.1",
    "cloudinary>=1.44.1",
]
